Develop

+++ database/migrations/2024_03_27_112137_create_formularios_table.php
@@ -5,6 +5,13 @@ use Illuminate\Database\Migrations\Migration;
 
 return new class extends Migration
 {
+	
+    /**
+     * Schema table name to migrate
+     * @var string
+     */
+    public $set_schema_table = 'formularios';
+
 	/**
 	 * Run the migrations.
 	 *
@@ -12,21 +19,29 @@ return new class extends Migration
 	 */
 	public function up()
 	{
-		Schema::create('formularios', function(Blueprint $table) {
+		if (Schema::hasTable($this->set_schema_table)) return;
+		Schema::create($this->set_schema_table, function(Blueprint $table) {
+            $table->engine = 'InnoDB';
             $table->increments('id');			
-            $table->unsignedInteger('setor_id');
-			$table->uuid('uuid');
-			$table->string('titulo')->nullable();
-			$table->string('sub_titulo')->nullable();
-			$table->integer('limite_caracteres')->nullable();
-			$table->date('ANO')->nullable();
+			$table->uuid('uuid')->unique();
+			$table->unsignedInteger('relatorio_id')->nullable();
+			$table->string('descricao')->nullable();
+            $table->integer('status')->default(0)->comment(
+				'
+					0 - Pendente,
+					1 - Não inciado,
+					2 - Concluído
+				'
+			);
+			$table->year('ano')->unique()->nullable();
 
             $table->timestamps();
             $table->softDeletes();
 
-			$table->index(["setor_id"], 'fk_formulario_setor_id_idx');
-            $table->foreign('setor_id','fk_formulario_setor_id_idx')
-                ->references('id')->on('teams')
+			$table->index(["relatorio_id"], 'fk_formulario_relatorio_idx');
+
+            $table->foreign('relatorio_id','fk_formulario_relatorio_idx')
+                ->references('id')->on('relatorios')
                 ->onDelete('restrict')
                 ->onUpdate('cascade');
 		});
@@ -39,7 +54,7 @@ return new class extends Migration
 	 */
 	public function down()
 	{
-		Schema::drop('formularios');
+		Schema::dropIfExists($this->set_schema_table);
 	}
 };
